using new staking sessionManage
